fixed CompatSharedExtensions
