+++ commons-core/src/main/scala-2.11/com/avsystem/commons/CompatSharedExtensions.scala
@@ -3,11 +3,14 @@ package com.avsystem.commons
 import com.avsystem.commons.concurrent.RunNowEC
 
 trait CompatSharedExtensions {
+
   import CompatSharedExtensions._
 
   implicit def futureCompatOps[A](fut: Future[A]): FutureCompatOps[A] = new FutureCompatOps(fut)
 
   implicit def futureCompanionCompatOps(fut: Future.type): FutureCompanionCompatOps.type = FutureCompanionCompatOps
+
+  implicit def tryCompatOps[A](tr: Try[A]): TryCompatOps[A] = new TryCompatOps(tr)
 }
 
 object CompatSharedExtensions {
@@ -34,5 +37,12 @@ object CompatSharedExtensions {
 
   object FutureCompanionCompatOps {
     final val unit: Future[Unit] = Future.successful(())
+  }
+
+  final class TryCompatOps[A](private val tr: Try[A]) extends AnyVal {
+    def fold[U](ft: Throwable => U, fa: A => U): U = tr match {
+      case Success(a) => fa(a)
+      case Failure(t) => ft(t)
+    }
   }
 }
